pyproject.toml
[project]
name = "relation-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "dspy>=3.0.4",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pyrootutils>=1.0.4",
    "scikit-learn>=1.7.2",
]
